requirements.txt
flask==3.1.0
gitpython==3.1.44
openai==1.66.2
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
